Tidy up type conversions

+++ src/genex/views2/slice.cpp
@@ -30,7 +30,7 @@ namespace genex::views {
         template <typename I, typename S, typename Int>
         requires detail::concepts::droppable_iters<I, S, Int> and std::random_access_iterator<I>
         GENEX_INLINE constexpr auto operator()(I first, S last, const Int n, const Int m) const {
-            return genex::span<iter_value_t<I>>(std::move(first) + n, std::move(first) + m);
+            return genex::span<iter_value_t<I>>(std::move(first) + n, std::move(first) + static_cast<std::size_t>(m));
         }
 
         template <typename I, typename S, typename Int>
@@ -43,7 +43,7 @@ namespace genex::views {
         requires detail::concepts::droppable_range<Rng, Int> and std::random_access_iterator<iterator_t<Rng>>
         GENEX_INLINE constexpr auto operator()(Rng &&rng, const Int n, const Int m) const {
             auto [first, last] = iterators::iter_pair(rng);
-            return genex::span<range_value_t<Rng>>(std::move(first) + n, std::move(first) + m);
+            return genex::span<range_value_t<Rng>>(std::move(first) + n, std::move(first) + static_cast<std::size_t>(m));
         }
 
         template <typename Rng, typename Int>
